refactor(lint): clear all eslint errors and warnings (#464)
* feat: lint and fixes

* format

* feat: precommit hook

* feat: conventional commit

* feat(infra): conventional commit errors msg

* refactor(lint): clear all eslint errors and warnings

Drives `yarn lint` from 96 errors + 14 warnings down to 0 + 0.

Highlights:
- eslint config: ignore `_`-prefixed unused args/vars/destructures.
- 22 explicit-any -> proper types (HoverOptions, PhaseIndicatorData,
  Record<id, pos>, MutableRefObject); 5 ternary statements -> if/else.
- 16 mixed-export files split into sibling .utils/.context modules
  (deckCover.utils, deckEditor.utils, deckBuilder.unsavedChanges,
  promptModalChrome.context); unused variant exports dropped.
- 18 setState-in-effect call sites converted to React docs' "previous
  value during render" pattern (no eslint-disable).
- react-hooks/refs + immutability + preserve-manual-memoization:
  dropped mirror refs in useBattlefieldLayout / HandDisplayNormal,
  switched FreeBattlefield/StackDisplay/PixiPhaseStripCanvas/
  DevToolsPanel to ResizeObserver state / useEffectEvent / singleton
  getter, refactored DeckListView positionsRef Map to useState,
  containerRef hoisted and passed via externalContainerRef.
- exhaustive-deps: real fixes (added missing deps, supplementaryCards
  and opponents wrapped in useMemo).

Verified: `yarn lint` clean, `tsc -b` clean, `vite dev` boots.
Pre-existing `vite build` failure due to stale wasm is unrelated.

* fix(lint): clear rust clippy errors and scope lint:rust

- forge-card-script: keys.contains() over iter().any() (2 sites).
- forge-foundation/card_type: collapse else-if.
- forge-foundation/mana: iter().find().copied() over manual loop.
- forge-server/connection: drop useless .into() calls.
- forge-server/error: box large WebSocket(tungstenite::Error) variant
  to fix result_large_err across 6 lobby.rs callers.
- forge-engine/static_ability_unspent_mana: real bug — for-loop that
  always returned on first item replaced with iter().find().
  clippy::never_loop is deny-by-default. Java short-circuit preserved.

Scope yarn lint:rust to --lib --bins (drops --all-targets / -D warnings).
Tests have pre-existing struct-field compile errors unrelated to lint;
the in-flight engine port carries clippy debt we are not addressing
in this PR. fmt check stays strict; clippy reports as warnings until
the port stabilises.

yarn lint:all now exits 0.

* chore(rust): clear most clippy warnings (880 -> 65)

Auto-fixes via `cargo clippy --fix` plus targeted manual fixes.
Highlights:

- Prefix unused trigger/effect params with `_` (host_card,
  host_controller, sa, ctx, runtime, state, etc).
- Replace `reader.lines().flatten()` with `map_while(Result::ok)` —
  real bug: flatten on Lines loops forever on read errors.
- forge-engine + forge-parity + forge-agent-interface +
  self-hosted-node + src-tauri: workspace-level allow for
  `clippy::too_many_arguments` (engine signatures mirror Java's
  multi-arg method shapes; refactoring would diverge from parity)
  and `clippy::type_complexity` / `clippy::doc_lazy_continuation`
  on forge-engine.

Remaining ~65 warnings are mostly in-flight dead code (port WIP) and
small style nits (identical if blocks for branching predicates that
return the same value, &PathBuf vs &Path, manual prefix strip, etc).
Leaving these for future cleanup as the engine port matures.

`yarn lint:all` exits 0; `yarn lint:rust` will re-enable -D warnings
once the port stabilises.

* chore(rust): clear remaining clippy warnings, enable -D warnings

Drives `yarn lint:rust` from ~880 + post-merge resurfaced warnings
down to zero, then re-enables `-D warnings` so future drift fails CI.

Bulk auto-fixes via `cargo clippy --fix` (needless borrows, redundant
closures, useless conversions, casts).

Targeted manual fixes:
- Prefix unused trigger params (`game`, `host_card`, `host_controller`)
  with `_` across ~145 trigger files; keep the originals where the
  body actually reads them.
- Replace manual `s[N..]` prefix/suffix slicing with `strip_prefix` /
  `strip_suffix` (cost_parser, player_property, valid_filter,
  ability_utils).
- Drop wildcard-with-redundant-pattern (`"X" | _`) in counters_put,
  phases_effect, attack_cost.
- `match`-as-equality collapsed into `if` in
  static_ability_alternative_cost.
- Replace `let Some(x) = ... else { return None; }` with `?` operator
  (mana::auto_pay).
- Remove `drop()` of non-Drop value in game_action; remove redundant
  shadow rebindings in action and cast_spell.
- Convert two `for _ in 0..n { vec.push(0); }` loops to
  `extend(repeat_n(0, n))`.
- Replace dead-code `let mut local_agents_storage = None;` overwrite
  pattern with explicit `#[allow(unused_assignments)]` justification.
- needless_range_loop refactors in deterministic_agent and main.

Workspace allows for in-flight engine / Java parity:
- forge-engine: too_many_arguments, type_complexity,
  doc_lazy_continuation, if_same_then_else (Java verbose conditional
  parity), collapsible_match.
- forge-agent-interface, forge-parity, self-hosted-node, src-tauri:
  too_many_arguments.

Per-item allows for genuinely intentional patterns: large_enum_variant
on AbilityValue / AgentPromptInner / RelayEnvelope; should_implement_trait
on ManaCostBeingPaid::next; only_used_in_recursion on
SpellAbility::set_card_state.

Per-item dead_code allows for in-flight scaffolding (build_mana_part,
matches_zone duplicates, valid_card_matches_with_source,
check_counter_qualifier, SKIP_PARAMS const, preset_decks helpers,
client_bot::bot_usernames, GameConfig.commander_name,
DeckList.cards, WasmGame.{human,ai}_player_id, Phase.phase_type).

Move `[profile.release]` from forge-wasm/Cargo.toml up to the
workspace Cargo.toml (kept `[package.metadata.wasm-pack.profile.release]`
in forge-wasm — that one stays at the package level by spec).

Real bugs removed along the way: never_loop in
static_ability_unspent_mana (loop body always returned on the first
iteration, masked the iter intent — replaced with iter().find()) and
flatten()-on-Lines in java_bridge (two sites, would loop forever on
read errors — replaced with map_while(Result::ok)).

Verified: `yarn lint:all` exits 0 with `-D warnings`,
`tsc -b` clean, `vite dev` boots in ~270ms.

* fix: test build

* feat: tauri start command

* refactor: rename XMage / bardidina references to OpenMagic

Type aliases and helpers used the legacy "XMage" branding; now align
with the OpenMagic product name.

- `XMageCard` -> `OpenMagicCard` across 21 files (101 sites). Imports
  alias `Card as OpenMagicCard` from `@/types/openmagic` to disambiguate
  from the shadcn `<Card>` UI component.
- `scryfallToXMage` -> `scryfallToOpenMagic`,
  `mapScryfallToXMage` -> `mapScryfallToOpenMagic`.
- "XMage Card" / "XMage card" -> "OpenMagic Card/card" in docstrings.
- `bardidina-deck-importer` User-Agent -> `openmagic-deck-importer` in
  archidekt.ts and moxfield.ts.

Also: route Tauri's beforeDev/Build commands through yarn so the
package manager matches the invoked one; eliminates the
`npm warn Unknown env config` lines that came from yarn's env vars
leaking into npm.

Left untouched: `GITHUB_REPO=fedepoi/bardidinaXmageUI` in
forge-server/.env — that's the literal GitHub repo URL, separate
rename concern.

+++ forge-engine/crates/forge-engine/src/ability/effects/change_zone_effect/search.rs
@@ -7,7 +7,6 @@ use forge_foundation::ZoneType;
 use super::super::{resolve_defined_players_with_sa, EffectContext};
 use super::helpers::{get_land_subtypes, matches_with_context};
 use crate::ids::{CardId, PlayerId};
-use crate::parsing::keys;
 use crate::spellability::SpellAbility;
 
 /// EACH clause search: one card per clause separated by "&".
@@ -183,7 +182,7 @@ fn resolve_constrained_multi(
     for _ in 0..max {
         // Apply budget filters
         if let Some(b) = budget_cmc {
-            remaining.retain(|&cid| ctx.game.card(cid).mana_cost.cmc() as i32 + spent_cmc <= b);
+            remaining.retain(|&cid| ctx.game.card(cid).mana_cost.cmc() + spent_cmc <= b);
         }
         if let Some(b) = budget_power {
             remaining.retain(|&cid| ctx.game.card(cid).base_power.unwrap_or(0) + spent_power <= b);
@@ -213,7 +212,7 @@ fn resolve_constrained_multi(
             required_land_types = land_types;
         }
 
-        spent_cmc += cmc as i32;
+        spent_cmc += cmc;
         spent_power += power;
         selected.push(chosen);
